Move Navigation class to strategies package
