+++ src/main/java/nl/tudelft/jpacman/strategies/Navigation.java
@@ -1,4 +1,4 @@
-package nl.tudelft.jpacman.npc.ghost;
+package nl.tudelft.jpacman.strategies;
 
 import nl.tudelft.jpacman.board.Board;
 import nl.tudelft.jpacman.board.Direction;
@@ -154,81 +154,5 @@ public final class Navigation {
             }
         }
         return null;
-    }
-
-    /**
-     * Helper class to keep track of the path.
-     *
-     * @author Jeroen Roosen
-     */
-    private static final class Node {
-
-        /**
-         * The direction for this node, which is <code>null</code> for the root
-         * node.
-         */
-        private final Direction direction;
-
-        /**
-         * The parent node, which is <code>null</code> for the root node.
-         */
-        private final Node parent;
-
-        /**
-         * The square associated with this node.
-         */
-        private final Square square;
-
-        /**
-         * Creates a new node.
-         *
-         * @param direction The direction, which is <code>null</code> for the root
-         *                  node.
-         * @param square    The square.
-         * @param parent    The parent node, which is <code>null</code> for the root
-         *                  node.
-         */
-        Node(Direction direction, Square square, Node parent) {
-            this.direction = direction;
-            this.square = square;
-            this.parent = parent;
-        }
-
-        /**
-         * @return The direction for this node, or <code>null</code> if this
-         * node is a root node.
-         */
-        private Direction getDirection() {
-            return direction;
-        }
-
-        /**
-         * @return The square for this node.
-         */
-        private Square getSquare() {
-            return square;
-        }
-
-        /**
-         * @return The parent node, or <code>null</code> if this node is a root
-         * node.
-         */
-        private Node getParent() {
-            return parent;
-        }
-
-        /**
-         * Returns the list of values from the root of the tree to this node.
-         *
-         * @return The list of values from the root of the tree to this node.
-         */
-        private List<Direction> getPath() {
-            if (parent == null) {
-                return new ArrayList<>();
-            }
-            List<Direction> path = parent.getPath();
-            path.add(getDirection());
-            return path;
-        }
     }
 }
